refactor: Rating 클래스의 패키지 위치 변경
