+++ src/main/java/success/planfit/repository/RatingRepository.java
@@ -1,7 +1,7 @@
 package success.planfit.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import success.planfit.entity.space.Rating;
+import success.planfit.entity.rating.Rating;
 
 public interface RatingRepository extends JpaRepository<Rating, Long> {
 }
